pyproject.toml
[project]
name = "lumina"
version = "0.1.0"
description = "Hindi-first AI assistant Lumina"
requires-python = ">=3.10"
dependencies = [
    "rich>=12.0.0",
    "pydantic>=1.10.0",
    "python-dotenv>=0.21.0",
    "httpx>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=0.991",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"